wip: initial Repository work [macata #2]
