+++ src/main/java/com/dazednconfused/catalauncher/database/mod/dao/ModH2DAOImpl.java
@@ -1,8 +1,8 @@
-package com.dazednconfused.catalauncher.database.mod;
+package com.dazednconfused.catalauncher.database.mod.dao;
 
 import com.dazednconfused.catalauncher.database.DAOException;
 import com.dazednconfused.catalauncher.database.H2Database;
-import com.dazednconfused.catalauncher.database.mod.entity.ModfileEntity;
+import com.dazednconfused.catalauncher.database.mod.entity.ModEntity;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -15,9 +15,11 @@ import java.util.Optional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class ModfileH2DAOImpl extends H2Database implements ModfileDAO {
+public class ModH2DAOImpl extends H2Database implements ModDAO {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(ModfileH2DAOImpl.class);
+    public static final String MODS_TABLE_NAME = "mod";
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(ModH2DAOImpl.class);
 
     @Override
     public String getDatabaseName() {
@@ -25,42 +27,39 @@ public class ModfileH2DAOImpl extends H2Database implements ModfileDAO {
     }
 
     /**
-     * Initializes the {@link #TABLE_NAME} table.
+     * Initializes the {@link #MODS_TABLE_NAME} table.
      * */
     @Override
     public Map.Entry<Connection, PreparedStatement> getTableCreationStatement() throws DAOException {
         try {
-            String sql = "CREATE TABLE IF NOT EXISTS " + TABLE_NAME + " (" +
+            String sql = "CREATE TABLE IF NOT EXISTS " + MODS_TABLE_NAME + " (" +
                 "id INT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY," +
-                "mod_id LONG NOT NULL, " +
-                "path TEXT NOT NULL," +
-                "hash TEXT NOT NULL," +
+                "name VARCHAR(255) NOT NULL," +
+                "modinfo TEXT NOT NULL," +
                 "created_date DATETIME NOT NULL," +
-                "updated_date DATETIME NOT NULL," +
-                "FOREIGN KEY (mod_id) REFERENCES " + ModDAO.TABLE_NAME + "(id)" +
+                "updated_date DATETIME NOT NULL" +
                 ")";
 
             Connection conn = this.getConnection();
             return Map.entry(conn, conn.prepareStatement(sql));
         } catch (SQLException e) {
-            LOGGER.error("An error occurred while crafting table creation statement for [{}]", TABLE_NAME, e);
+            LOGGER.error("An error occurred while crafting table creation statement for [{}]", MODS_TABLE_NAME, e);
             throw new DAOException(e);
         }
     }
 
     @Override
-    public ModfileEntity insert(ModfileEntity entity) throws DAOException {
-        LOGGER.debug("Inserting ModfileEntity [{}]...", entity);
+    public ModEntity insert(ModEntity entity) throws DAOException {
+        LOGGER.debug("Inserting ModEntity [{}]...", entity);
 
-        String sql = "INSERT INTO " + TABLE_NAME + "" +
-            "(mod_id, path, hash, created_date, updated_date) " +
+        String sql = "INSERT INTO " + MODS_TABLE_NAME + "" +
+            "(name, modinfo, created_date, updated_date) " +
             "VALUES " +
-            "(?, ?, ?, CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP())";
+            "(?, ?, CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP())";
 
         try (Connection conn = this.getConnection(); PreparedStatement pstmt = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
-            pstmt.setLong(1, entity.getModId());
-            pstmt.setString(2, entity.getPath());
-            pstmt.setString(3, entity.getHash());
+            pstmt.setString(1, entity.getName());
+            pstmt.setString(2, entity.getModinfo());
 
             pstmt.executeUpdate();
 
@@ -72,24 +71,24 @@ public class ModfileH2DAOImpl extends H2Database implements ModfileDAO {
     }
 
     @Override
-    public ModfileEntity update(ModfileEntity entity) throws DAOException {
-        Optional<ModfileEntity> originalEntity = this.findById(entity.getId());
+    public ModEntity update(ModEntity entity) throws DAOException {
+        Optional<ModEntity> originalEntity = this.findById(entity.getId());
 
         if (originalEntity.isEmpty()) {
             throw new DAOException("No entity with id [" + entity.getId() + "] found");
         }
 
-        LOGGER.debug("Updating ModfileEntity from [{}] to [{}]...", originalEntity.get(), entity);
+        LOGGER.debug("Updating ModEntity from [{}] to [{}]...", originalEntity.get(), entity);
 
-        String sql = "UPDATE " + TABLE_NAME + " SET " +
-            "path = ?, " +
-            "hash = ?, " +
+        String sql = "UPDATE " + MODS_TABLE_NAME + " SET " +
+            "name = ?, " +
+            "modinfo = ?, " +
             "updated_date = CURRENT_TIMESTAMP() " +
             "WHERE id = ?";
 
         try (Connection conn = this.getConnection(); PreparedStatement pstmt = conn.prepareStatement(sql)) {
-            pstmt.setString(1, entity.getPath());
-            pstmt.setString(2, entity.getHash());
+            pstmt.setString(1, entity.getName());
+            pstmt.setString(2, entity.getModinfo());
             pstmt.setLong(3, entity.getId());
 
             pstmt.executeUpdate();
@@ -102,14 +101,14 @@ public class ModfileH2DAOImpl extends H2Database implements ModfileDAO {
     }
 
     @Override
-    public ModfileEntity buildFromResultSet(ResultSet rs) throws DAOException {
-        LOGGER.trace("Building ModfileEntity from ResultSet [{}]...", rs);
+    public ModEntity buildFromResultSet(ResultSet rs) throws DAOException {
+        LOGGER.trace("Building ModEntity from ResultSet [{}]...", rs);
 
         try {
-            return ModfileEntity.builder()
+            return ModEntity.builder()
                 .id(rs.getLong("id"))
-                .path(rs.getString("path"))
-                .hash(rs.getString("hash"))
+                .name(rs.getString("name"))
+                .modinfo(rs.getString("modinfo"))
                 .createdDate(rs.getTimestamp("created_date"))
                 .updatedDate(rs.getTimestamp("updated_date"))
                 .build();
