+++ src/main/java/com/dazednconfused/catalauncher/database/mod/dao/ModfileH2DAOImpl.java
@@ -1,8 +1,8 @@
-package com.dazednconfused.catalauncher.database.mod;
+package com.dazednconfused.catalauncher.database.mod.dao;
 
 import com.dazednconfused.catalauncher.database.DAOException;
 import com.dazednconfused.catalauncher.database.H2Database;
-import com.dazednconfused.catalauncher.database.mod.entity.ModEntity;
+import com.dazednconfused.catalauncher.database.mod.entity.ModfileEntity;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -15,13 +15,9 @@ import java.util.Optional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class ModH2DAOImpl extends H2Database implements ModDAO {
+public class ModfileH2DAOImpl extends H2Database implements ModfileDAO {
 
-    public static final String MODS_TABLE_NAME = "mod";
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(ModH2DAOImpl.class);
-
-    private static final String DATABASE_FILE = "mods";
+    private static final Logger LOGGER = LoggerFactory.getLogger(ModfileH2DAOImpl.class);
 
     @Override
     public String getDatabaseName() {
@@ -29,39 +25,42 @@ public class ModH2DAOImpl extends H2Database implements ModDAO {
     }
 
     /**
-     * Initializes the {@link #MODS_TABLE_NAME} table.
+     * Initializes the {@link #TABLE_NAME} table.
      * */
     @Override
     public Map.Entry<Connection, PreparedStatement> getTableCreationStatement() throws DAOException {
         try {
-            String sql = "CREATE TABLE IF NOT EXISTS " + MODS_TABLE_NAME + " (" +
+            String sql = "CREATE TABLE IF NOT EXISTS " + TABLE_NAME + " (" +
                 "id INT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY," +
-                "name VARCHAR(255) NOT NULL," +
-                "modinfo TEXT NOT NULL," +
+                "mod_id LONG NOT NULL, " +
+                "path TEXT NOT NULL," +
+                "hash TEXT NOT NULL," +
                 "created_date DATETIME NOT NULL," +
-                "updated_date DATETIME NOT NULL" +
+                "updated_date DATETIME NOT NULL," +
+                "FOREIGN KEY (mod_id) REFERENCES " + ModDAO.TABLE_NAME + "(id)" +
                 ")";
 
             Connection conn = this.getConnection();
             return Map.entry(conn, conn.prepareStatement(sql));
         } catch (SQLException e) {
-            LOGGER.error("An error occurred while crafting table creation statement for [{}]", MODS_TABLE_NAME, e);
+            LOGGER.error("An error occurred while crafting table creation statement for [{}]", TABLE_NAME, e);
             throw new DAOException(e);
         }
     }
 
     @Override
-    public ModEntity insert(ModEntity entity) throws DAOException {
-        LOGGER.debug("Inserting ModEntity [{}]...", entity);
+    public ModfileEntity insert(ModfileEntity entity) throws DAOException {
+        LOGGER.debug("Inserting ModfileEntity [{}]...", entity);
 
-        String sql = "INSERT INTO " + MODS_TABLE_NAME + "" +
-            "(name, modinfo, created_date, updated_date) " +
+        String sql = "INSERT INTO " + TABLE_NAME + "" +
+            "(mod_id, path, hash, created_date, updated_date) " +
             "VALUES " +
-            "(?, ?, CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP())";
+            "(?, ?, ?, CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP())";
 
         try (Connection conn = this.getConnection(); PreparedStatement pstmt = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
-            pstmt.setString(1, entity.getName());
-            pstmt.setString(2, entity.getModinfo());
+            pstmt.setLong(1, entity.getModId());
+            pstmt.setString(2, entity.getPath());
+            pstmt.setString(3, entity.getHash());
 
             pstmt.executeUpdate();
 
@@ -73,24 +72,24 @@ public class ModH2DAOImpl extends H2Database implements ModDAO {
     }
 
     @Override
-    public ModEntity update(ModEntity entity) throws DAOException {
-        Optional<ModEntity> originalEntity = this.findById(entity.getId());
+    public ModfileEntity update(ModfileEntity entity) throws DAOException {
+        Optional<ModfileEntity> originalEntity = this.findById(entity.getId());
 
         if (originalEntity.isEmpty()) {
             throw new DAOException("No entity with id [" + entity.getId() + "] found");
         }
 
-        LOGGER.debug("Updating ModEntity from [{}] to [{}]...", originalEntity.get(), entity);
+        LOGGER.debug("Updating ModfileEntity from [{}] to [{}]...", originalEntity.get(), entity);
 
-        String sql = "UPDATE " + MODS_TABLE_NAME + " SET " +
-            "name = ?, " +
-            "modinfo = ?, " +
+        String sql = "UPDATE " + TABLE_NAME + " SET " +
+            "path = ?, " +
+            "hash = ?, " +
             "updated_date = CURRENT_TIMESTAMP() " +
             "WHERE id = ?";
 
         try (Connection conn = this.getConnection(); PreparedStatement pstmt = conn.prepareStatement(sql)) {
-            pstmt.setString(1, entity.getName());
-            pstmt.setString(2, entity.getModinfo());
+            pstmt.setString(1, entity.getPath());
+            pstmt.setString(2, entity.getHash());
             pstmt.setLong(3, entity.getId());
 
             pstmt.executeUpdate();
@@ -103,14 +102,14 @@ public class ModH2DAOImpl extends H2Database implements ModDAO {
     }
 
     @Override
-    public ModEntity buildFromResultSet(ResultSet rs) throws DAOException {
-        LOGGER.trace("Building ModEntity from ResultSet [{}]...", rs);
+    public ModfileEntity buildFromResultSet(ResultSet rs) throws DAOException {
+        LOGGER.trace("Building ModfileEntity from ResultSet [{}]...", rs);
 
         try {
-            return ModEntity.builder()
+            return ModfileEntity.builder()
                 .id(rs.getLong("id"))
-                .name(rs.getString("name"))
-                .modinfo(rs.getString("modinfo"))
+                .path(rs.getString("path"))
+                .hash(rs.getString("hash"))
                 .createdDate(rs.getTimestamp("created_date"))
                 .updatedDate(rs.getTimestamp("updated_date"))
                 .build();
