+++ src/test/java/com/dazednconfused/catalauncher/database/mod/dao/ModDAOTest.java
@@ -1,7 +1,6 @@
-package com.dazednconfused.catalauncher.database.mod;
+package com.dazednconfused.catalauncher.database.mod.dao;
 
 import com.dazednconfused.catalauncher.database.BaseDAO;
-import com.dazednconfused.catalauncher.database.H2Database;
 import com.dazednconfused.catalauncher.database.mod.entity.ModEntity;
 
 import org.junit.jupiter.api.*;
